requirements.txt
google-generativeai
python-dotenv
numpy
opencv-python
fastapi
uvicorn
sqlalchemy
python-multipart